+++ Scripts/run-benchmark-history.ps1
@@ -41,7 +41,7 @@ function RestoreBenchmark() {
     Invoke-Expression "sed -i 's/\\Performance.Tests.csproj/\\Microsoft.Coyote.Performance.Tests.csproj/' $RootDir\Coyote.sln"
 }
 
-$benchmarks_dir = "$RootDir/Tools/BenchmarkRunner/bin/net6.0"
+$benchmarks_dir = "$RootDir/Tools/BenchmarkRunner/bin/net7.0"
 $benchmark_runner = "BenchmarkRunner.exe"
 $index = 0